Add ChefSpec matcher, correct deprecation warning
Added overdue ChefSpec matcher (shout out to @chazzly for the poke).
Added using PoiseHelpers to make it match the rest of the Poise/Halite
pattern.

Also added a fix for the TrueClass deprecation warning where false as a
default to a TrueClass parameter will actually trigger an error in Chef
13.